Printf

+++ main.h
@@ -11,7 +11,7 @@ int _printf(const char *format, ...);
 
 int process_format(const char *format, va_list args);
 int process_str(const char *str);
-long print_number(long num);
-int process_int(va_list args);
+long int print_number(long int num);
+long int process_int(va_list args);
 
 #endif
